Fix bot not resuming after shard reconnect
Still testing this, there are still a ton of debug comments in the code and console output but it so far seems to be working. I'll leave it running for a few more days just to make sure it still works and it doesn't yeet half the database when it feels like it.